split wf/wft endpoints and add get wft
