+++ fractal_server/app/api/__init__.py
@@ -10,6 +10,7 @@ from .v1.job import router as job_router
 from .v1.project import router as project_router
 from .v1.task import router as task_router
 from .v1.workflow import router as workflow_router
+from .v1.workflowtask import router as workflowtask_router
 
 
 router_default = APIRouter()
@@ -19,6 +20,7 @@ router_v1.include_router(project_router, prefix="/project", tags=["Projects"])
 router_v1.include_router(task_router, prefix="/task", tags=["Tasks"])
 router_v1.include_router(dataset_router, tags=["Datasets"])
 router_v1.include_router(workflow_router, tags=["Workflows"])
+router_v1.include_router(workflowtask_router, tags=["Workflow Tasks"])
 router_v1.include_router(job_router, tags=["Jobs"])
